Fixing renaming issues
