--- src/SFA.DAS.LevyTransferMatching.Api/Models/Pledges/GetMyPledgesResponse.cs
@@ -1,17 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace SFA.DAS.LevyTransferMatching.Api.Models.Pledges
-{
-    public class GetMyPledgesResponse
-    {
-        public IEnumerable<MyPledge> Pledges { get; set; }
-
-        public class MyPledge
-        {
-            public int Id { get; set; }
-            public int Amount { get; set; }
-            public int RemainingAmount { get; set; }
-            public int ApplicationCount { get; set; }
-        }
-    }
-}
