+++ src/SFA.DAS.LevyTransferMatching/Application/Queries/Pledges/GetPledges/GetPledgesQueryResult.cs
@@ -4,9 +4,9 @@ namespace SFA.DAS.LevyTransferMatching.Application.Queries.Pledges.GetPledges
 {
     public class GetPledgesQueryResult
     {
-        public IEnumerable<MyPledge> Pledges { get; set; }
+        public IEnumerable<Pledge> Pledges { get; set; }
 
-        public class MyPledge
+        public class Pledge
         {
             public int Id { get; set; }
             public int Amount { get; set; }
